# api/openapi-spec/openapi.yaml
openapi: 3.0.0
info:
  title: Optimizely Agent API
  description: Optimizely Agent - Full Stack REST API
  termsOfService: http://optimizely.com/terms/
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.12.0
security:
  - SdkKeyAuth: []
  - TokenAuth: []
paths:
...
  /v1/override:
    post:
      summary: Override an experiment decision for a user
      operationId: override
      description: Override a experiment or feature test decision used in future user based decisions. This override is only stored locally
        in memory for debugging and testing purposes and should not be used
        for production overrides.
      responses:
        '201':
          description: Forced variation set
        '204':
          description: Forced variation was already set
        '400':
          description: Invalid user id
        '401':
          description: Unauthorized, invalid JWT
        '403':
          $ref: '#/components/responses/Forbidden'
      requestBody:
        $ref: '#/components/requestBodies/OverrideContext'
...
components:
...
  requestBodies:
...
    OverrideContext:
      required: true
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/OverrideContext'
...
  responses:
    Forbidden:
      description: You do not have necessary permissions for the resource
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      properties:
        error:
          type: string
...
    OverrideContext:
      type: array
      items:
        type: object
        properties:
          userId:
            type: string
          experimentKey:
            type: string
          variationKey:
            type: string
...
  securitySchemes:
    SdkKeyAuth:
      in: header
      name: X-Optimizely-SDK-Key
      type: apiKey
    TokenAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT